Refactor Player Controller
Extract player/camera movement into its own class
Extract Animation into its own class
Extract Collectables methods into its own class
Extract UI update methods into its own class
