+++ Assets/Scripts/CollectableController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b8c21d1cace1b34ba5e21df1b4f2691
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
