+++ Assets/Scripts/PlayerAnimate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b4f5143682c3af84bbd35e9fbb9a5622
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
