+++ Assets/Scripts/UpdateUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 815d7d4f457407e47a045e31a38d290f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
